fix(build): repair the test targets and stop tracking build artifacts
"make test" could not run. A backslash sat mid-line in ERROR_PAGE_TEST_SRC,
so make took the backslash for a prerequisite and tried to build a file
named ".o"; the error_page and multipart targets were also missing sources
they link against, and server_config_test compared a size_t against -1L,
which -Werror rejects.

Ten compiled test binaries and three symlinks into /tmp/evil_outside were
committed alongside them. The binaries are x86-64 ELF, so make found them
newer than their sources and ran them unbuilt, which fails on any other
architecture. The symlinks are leftovers of the symlink-escape test and
pointed out of the served directory. Both are dropped and ignored.

With this the suite builds and runs: 173 assertions, 0 failures.

+++ tests/server_config_test.cpp
@@ -104,7 +104,7 @@ static void	test_location_config_defaults(void)
 	         "default location has no method restriction");
 	CHECK_EQ(location.returnCode, 0,
 	         "default location has no return configured");
-	CHECK_EQ(location.clientMaxBodySize, -1L,
+	CHECK_EQ(location.clientMaxBodySize, static_cast<size_t>(-1),
 	         "default location inherits client_max_body_size from server");
 }
